Rename and move transform func into something more specific
